requirements.txt
# StrokeSense - AI-Powered Stroke Prediction System
# Python Dependencies

# ============================================
# Web Framework (Backend API)
# ============================================
flask>=3.0.0
flask-cors>=4.0.0

# ============================================
# Deep Learning
# ============================================
tensorflow>=2.13.0

# ============================================
# Machine Learning
# ============================================
scikit-learn>=1.3.0
joblib>=1.3.0
imbalanced-learn>=0.11.0

# ============================================
# Data Processing
# ============================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ============================================
# Image Processing
# ============================================
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# ============================================
# Visualization
# ============================================
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# ============================================
# Streamlit (Legacy Frontend)
# ============================================
streamlit>=1.28.0

# ============================================
# Utilities
# ============================================
python-dateutil>=2.8.0
